pyproject.toml
[project]
name = "codepilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.24.0",
    "keyring>=25.7.0",
    "litellm>=1.81.11",
    "python-dotenv>=1.2.1",
    "questionary>=2.1.1",
    "rich>=14.3.2",
    "typer>=0.23.1",
]

[project.scripts]
codepilot = "app.cli.main:app"
